+++ src/main/kotlin/com/rankweis/uppercut/karate/psi/element/KarateNamedElement.java
@@ -1,0 +1,7 @@
+package com.rankweis.uppercut.karate.psi.element;
+
+import com.intellij.psi.PsiNameIdentifierOwner;
+
+public interface KarateNamedElement extends PsiNameIdentifierOwner {
+
+}