+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 # uppercut Changelog
 
-## [0.0.3]
+## [0.0.5]
+### Added
+- Some better syntax highlighting/fixes
+
+### Removed
+- Dependency on karate-junit and now just has karate-core, should be much smaller.
+
+## [0.0.4]
 ### Added
 - Dependency on karate 1.5.0.RC3 for ease of development